Setup basic server and client architecture
