+++ client/.env
@@ -1,0 +1,2 @@
+# These are public and will be injected into the build
+REACT_APP_API_BASE_URL=http://localhost:4000
