+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "formpro",
+  "name": "formpro-client",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -7,13 +7,8 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.4.0",
-    "body-parser": "^1.20.2",
     "bootstrap": "^5.3.0",
-    "express": "^4.18.2",
     "google-libphonenumber": "^3.2.32",
-    "mongodb": "^5.6.0",
-    "mongoose": "^7.3.0",
-    "mongosh": "^1.10.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
     "react-dom": "^18.2.0",
