+++ client/src/App.js
@@ -1,0 +1,19 @@
+import "./App.css";
+import Forrm from "./Components/Forrm";
+import Forrmd from "./Components/Forrmd";
+import Head1 from "./Components/Head1";
+import Messages from "./Components/Messages";
+import Parllax from "./Components/Parllax";
+
+function App() {
+  return (
+    <>
+      {/* <Head1 /> */}
+      {/* <Parllax /> */}
+      {/* <Forrmd/> */}
+      <Messages />
+    </>
+  );
+}
+
+export default App;
