+++ client/src/Components/Popcontent.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Popcontent() {
+  return (
+    <div>Popcontent</div>
+  )
+}
+
+export default Popcontent
