+++ client/src/constants.js
@@ -1,0 +1,2 @@
+// Environment variables
+export const API_BASE_URL = process.env.REACT_APP_API_BASE_URL
