+++ client/src/utils/api.client.js
@@ -1,0 +1,6 @@
+import axios from 'axios'
+import { API_BASE_URL } from '../constants'
+
+export const api = axios.create({
+  baseURL: API_BASE_URL
+})
