+++ server/.env.example
@@ -1,0 +1,3 @@
+# Use this file as an example for your own .env file
+PORT=4000
+MONGODB_URI=mongodb://localhost:27017/anxietyform
