--- src/App.js
@@ -1,19 +1,0 @@
-import './App.css';
-import Forrm from './Components/Forrm';
-import Forrmd from './Components/Forrmd';
-import Head1 from './Components/Head1';
-import Parllax from './Components/Parllax';
-
-
-function App() {
-  return (
-    <>
-     <Head1/>
-     <Parllax/>
-     <Forrm/>
-    {/* <Forrmd/> */}
-    </>
-  );
-}
-
-export default App;
