--- src/Components/Popcontent.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Popcontent() {
-  return (
-    <div>Popcontent</div>
-  )
-}
-
-export default Popcontent
