--- src/db.js
@@ -1,23 +1,0 @@
-// const { MongoClient } = require('mongodb');
-
-// const uri = 'mongodb+srv://harmeeths1998:harmeeths1998@cluster0.7qtb3ih.mongodb.net/?retryWrites=true&w=majority'; // Replace with your MongoDB Atlas connection string
-
-// let client;
-
-// async function connect() {
-//   try {
-//     client = await MongoClient.connect(uri);
-//     console.log('Connected to MongoDB Atlas');
-//   } catch (error) {
-//     console.error('Error connecting to MongoDB Atlas:', error);
-//   }
-// }
-
-// function getClient() {
-//   return client;
-// }
-
-// module.exports = {
-//   connect,
-//   getClient,
-// };
